feat: open terminal in adjacent window on large screens Extract multi-window intent flags and handle new intents in TerminalActivity.
